+++ .coveragerc
@@ -1,7 +1,7 @@
 [run]
 omit =
     predicators/envs/behavior.py
-    predicators/envs/behavior_options.py
+    predicators/behavior_utils/**
     predicators/third_party/**
 
 [report]
