+++ predicators/envs/behavior.py
@@ -36,11 +36,14 @@ except (ImportError, ModuleNotFoundError) as e:
 from gym.spaces import Box
 
 from predicators import utils
+from predicators.behavior_utils.option_fns import create_grasp_policy, \
+    create_navigate_policy, create_place_policy
+from predicators.behavior_utils.option_model_fns import \
+    create_grasp_option_model, create_navigate_option_model, \
+    create_place_option_model
+from predicators.behavior_utils.planner_fns import make_grasp_plan, \
+    make_navigation_plan, make_placeontop_plan
 from predicators.envs import BaseEnv
-from predicators.envs.behavior_options import create_grasp_option_model, \
-    create_grasp_policy, create_navigate_option_model, \
-    create_navigate_policy, create_place_option_model, create_place_policy, \
-    grasp_obj_at_pos, navigate_to_obj_pos, place_ontop_obj_pos
 from predicators.settings import CFG
 from predicators.structs import Action, Array, GroundAtom, Object, \
     ParameterizedOption, Predicate, State, Task, Type, Video
@@ -81,7 +84,7 @@ class BehaviorEnv(BaseEnv):
             "behavior_env.BehaviorEnv", Union[
                 "URDFObject", "RoomFloor"], Array, Optional[Generator]
         ], Optional[Tuple[List[List[float]], List[List[float]]]]]] = [
-            navigate_to_obj_pos, grasp_obj_at_pos, place_ontop_obj_pos
+            make_navigation_plan, make_grasp_plan, make_placeontop_plan
         ]
         option_policy_fns: List[
             Callable[[List[List[float]], List[List[float]]],
