+++ tests/behavior_utils/test_behavior_utils.py
@@ -1,3 +1,5 @@
+"""Tests for behavior_utils."""
+
 import numpy as np
 import pytest
